move associator to its own module
